Move helpers to misc_test.go
